# Compare fields directly instead of PackedStyle