@jcollie 2025-09-02
Curious if the CPU architecture makes a difference (aarch64 vs x86_64).