@j-c-m 2025-09-02
Here is updated benchmark using the comptime coverage of Style struct.

## M2 aarch64

Style.eql Benchmark Comparison
==============================

Test: Small (1K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 58205
  Duration: 500.00 ms
  Throughput: 116409069 comparisons/sec
Old implementation:
  Iterations: 8528
  Duration: 500.00 ms
  Throughput: 17055932 comparisons/sec
Performance improvement:
  Speedup: 6.83x
  Improvement: +582.5%

Test: Medium (10K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 4667
  Duration: 500.04 ms
  Throughput: 93332160 comparisons/sec
Old implementation:
  Iterations: 852
  Duration: 500.39 ms
  Throughput: 17026855 comparisons/sec
Performance improvement:
  Speedup: 5.48x
  Improvement: +448.1%

Test: Large (50K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 922
  Duration: 500.17 ms
  Throughput: 92168478 comparisons/sec
Old implementation:
  Iterations: 171
  Duration: 502.69 ms
  Throughput: 17008427 comparisons/sec
Performance improvement:
  Speedup: 5.42x
  Improvement: +441.9%

Test: Mostly equal (10K pairs, 90% equal)
--------------------------------------------------
New implementation:
  Iterations: 4968
  Duration: 500.07 ms
  Throughput: 99345893 comparisons/sec
Old implementation:
  Iterations: 851
  Duration: 500.17 ms
  Throughput: 17014079 comparisons/sec
Performance improvement:
  Speedup: 5.84x
  Improvement: +483.9%

Test: Mostly different (10K pairs, 10% equal)
--------------------------------------------------
New implementation:
  Iterations: 4314
  Duration: 500.06 ms
  Throughput: 86270165 comparisons/sec
Old implementation:
  Iterations: 851
  Duration: 500.58 ms
  Throughput: 17000212 comparisons/sec
Performance improvement:
  Speedup: 5.07x
  Improvement: +407.5%

Test: Same flags (10K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 2881
  Duration: 500.06 ms
  Throughput: 57613662 comparisons/sec
Old implementation:
  Iterations: 861
  Duration: 500.22 ms
  Throughput: 17212358 comparisons/sec
Performance improvement:
  Speedup: 3.35x
  Improvement: +234.7%

## Slower old FreeBSD x86_64 machine (what I had available)

Test: Small (1K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 6210
  Duration: 500.02 ms
  Throughput: 12419432 comparisons/sec
Old implementation:
  Iterations: 1584
  Duration: 500.01 ms
  Throughput: 3167964 comparisons/sec
Performance improvement:
  Speedup: 3.92x
  Improvement: +292.0%

Test: Medium (10K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 625
  Duration: 500.30 ms
  Throughput: 12492523 comparisons/sec
Old implementation:
  Iterations: 158
  Duration: 501.36 ms
  Throughput: 3151399 comparisons/sec
Performance improvement:
  Speedup: 3.96x
  Improvement: +296.4%

Test: Large (50K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 126
  Duration: 503.70 ms
  Throughput: 12507559 comparisons/sec
Old implementation:
  Iterations: 32
  Duration: 507.51 ms
  Throughput: 3152663 comparisons/sec
Performance improvement:
  Speedup: 3.97x
  Improvement: +296.7%

Test: Mostly equal (10K pairs, 90% equal)
--------------------------------------------------
New implementation:
  Iterations: 634
  Duration: 500.40 ms
  Throughput: 12669788 comparisons/sec
Old implementation:
  Iterations: 159
  Duration: 503.15 ms
  Throughput: 3160118 comparisons/sec
Performance improvement:
  Speedup: 4.01x
  Improvement: +300.9%

Test: Mostly different (10K pairs, 10% equal)
--------------------------------------------------
New implementation:
  Iterations: 616
  Duration: 500.14 ms
  Throughput: 12316592 comparisons/sec
Old implementation:
  Iterations: 158
  Duration: 501.77 ms
  Throughput: 3148875 comparisons/sec
Performance improvement:
  Speedup: 3.91x
  Improvement: +291.1%

Test: Same flags (10K pairs, 50% equal)
--------------------------------------------------
New implementation:
  Iterations: 537
  Duration: 500.56 ms
  Throughput: 10727918 comparisons/sec
Old implementation:
  Iterations: 159
  Duration: 502.91 ms
  Throughput: 3161580 comparisons/sec
Performance improvement:
  Speedup: 3.39x
  Improvement: +239.3%