@jcollie 2025-09-02
It would be good to include the benchmarking code so that others can verify them, and so that we can run them in the future when we make major changes (like the upcoming Zig 0.15 transition).